split out pipeline definition initialization for csv and lemonade pipelines
